fix(rust): reorganize code, address clippy errors
