+++ implementations/rust/daemon/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cli;
+pub mod config;
 pub mod initiator;
 pub mod node;
 pub mod responder;
