+++ implementations/rust/daemon/src/responder.rs
@@ -1,6 +1,7 @@
 use std::io::Write;
 
-use crate::node::{Config, Node};
+use crate::config::Config;
+use crate::node::Node;
 use crate::worker::Worker;
 
 use ockam_message::message::RouterAddress;
@@ -9,9 +10,9 @@ pub fn run(config: Config) {
     let (mut node, router_tx) = Node::new(&config);
 
     let worker_addr = RouterAddress::worker_router_address_from_str("01242020").unwrap();
-    let worker = Worker::new(worker_addr, router_tx.clone(), |_self, msg| {
+    let worker = Worker::new(worker_addr, router_tx, |_self, msg| {
         let mut out = std::io::stdout();
-        out.write(msg.message_body.as_ref())
+        out.write_all(msg.message_body.as_ref())
             .expect("failed to write message to stdout");
         out.flush().expect("failed to flush stdout");
     });
